+++ java/javaapi/src/test/java/joynr/test/NonInstantiable.java
@@ -1,7 +1,7 @@
 /*
  * #%L
  * %%
- * Copyright (C) 2017 BMW Car IT GmbH
+ * Copyright (C) 2018 BMW Car IT GmbH
  * %%
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,27 +16,8 @@
  * limitations under the License.
  * #L%
  */
-package io.joynr.provider;
+package joynr.test;
 
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Inherited;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-/**
- *
- * @author christoph.ainhauser
- */
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-@Inherited
-public @interface JoynrInterface {
-
-    Class<?> provides();
-
-    Class<?> provider();
-
-    String name();
+public interface NonInstantiable {
 
 }