+++ javascript/libjoynr-js/src/test/js/node_integration/ChildProcessUtils.js
@@ -52,7 +52,7 @@ ChildProcessUtils.registerHandlers = function(initializeTest, startTest, termina
             if (!initializeTest) {
                 throw new Error("cannot initialize test, child does not define an initializeTest method");
             }
-            initializeTest(msg.provisioningSuffix, msg.domain).then(providedRuntime => {
+            initializeTest(msg.provisioningSuffix, msg.domain, msg.processSpecialization).then(providedRuntime => {
                 runtime = providedRuntime;
                 ChildProcessUtils.postReady();
             });