+++ tools/generator/cpp-generator/src/test/java/io/joynr/generator/cpp/util/CppStdTypeUtilTest.java
@@ -20,6 +20,7 @@ package io.joynr.generator.cpp.util;
 
 import io.joynr.generator.loading.ModelLoader;
 import io.joynr.generator.templates.util.JoynrGeneratorExtensions;
+import io.joynr.generator.templates.util.NamingUtil;
 
 import java.net.URL;
 
@@ -56,6 +57,10 @@ public class CppStdTypeUtilTest extends TestCase {
                 bindConstant().annotatedWith(Names.named("generationId")).to("");
                 bindConstant().annotatedWith(Names.named(JoynrGeneratorExtensions.JOYNR_GENERATOR_CLEAN)).to(false);
                 bindConstant().annotatedWith(Names.named(JoynrGeneratorExtensions.JOYNR_GENERATOR_GENERATE)).to(true);
+                bindConstant().annotatedWith(Names.named(JoynrGeneratorExtensions.JOYNR_GENERATOR_PACKAGEWITHVERSION))
+                              .to(false);
+                bindConstant().annotatedWith(Names.named(NamingUtil.JOYNR_GENERATOR_INTERFACENAMEWITHVERSION))
+                              .to(false);
                 install(new FactoryModuleBuilder().build(CppTemplateFactory.class));
             }
         }).getInstance(CppStdTypeUtil.class);